+++ packages/node/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Update Stellar SDK to support protocol 23 (#135)
+- Update Stellar SDK to v14 with support for protocol 23 (#138)
+
+### Fixed
+- Pagination for getting events which used the deprecated pagingToken, it now uses the cursor (#138)
 
 ## [6.0.2] - 2025-07-21
 ### Fixed